requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
O365==2.0.21
oauthlib==3.2.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-gitlab==3.11.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
six==1.16.0
soupsieve==2.3.2.post1
stringcase==1.2.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
